# app/Http/Resources/ShopResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

/**
 *
 * @OA\Schema(
 * @OA\Xml(name="Shops"),
 * @OA\Property(property="id", type="integer", readOnly="true", example="1"),
 * @OA\Property(property="name", type="string", maxLength=32, example="Digitec"),
 * @OA\Property(property="link", type="string", maxLength=32, example="www.digitec.ch"),
 * @OA\Property(property="products", type="array", collectionFormat="multi",
 *      @OA\Items(ref="#/components/schemas/ProductResource"),
 * ),
 * )
 *
 */

class ShopResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @return array<string, mixed>
     */
    public function toArray(Request $request): array
    {
        return [
            'id' => $this->id,
            'name' => $this->name,
            'link' => $this->link,
            'products' => ProductResource::collection($this->products)
        ];
    }
}