# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bestellungs-Übersicht | API-Backend in Laravel",
        "description": "Diese Web-App verwaltet Bestellungen und die dazugehörenden Produkte und Geschäfte.\n *     Dieses Projekt wurde im Laufe einer IPA erstellt.",
        "contact": {
            "email": "julien.raedler@twofold.swiss"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/shops": {
            "get": {
                "tags": [
                    "shop"
                ],
                "summary": "Get all stores",
                "description": "Returns list of stores",
                "operationId": "getStores",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "shops": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ShopResource"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "shop"
                ],
                "summary": "Create new store",
                "description": "Creates one new store",
                "operationId": "storeStore",
                "requestBody": {
                    "description": "Enter shop Data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "link"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Digitec"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "www.digitec.ch"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "link"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Digitec"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "www.digitec.ch"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "shop": {
                                            "$ref": "#/components/schemas/ShopResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shops/{shopId}": {
            "get": {
                "tags": [
                    "shop"
                ],
                "summary": "Show one store",
                "description": "Get one store by it's id",
                "operationId": "showStore",
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "description": "ID of shop",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "shop": {
                                            "$ref": "#/components/schemas/ShopResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "shop"
                ],
                "summary": "Update a store",
                "description": "Update the parameters of a store",
                "operationId": "updateStore",
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "description": "ID of shop",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Enter shop Data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Digitec"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "www.digitec.ch"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "shop": {
                                            "$ref": "#/components/schemas/ShopResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "shop"
                ],
                "summary": "Delete one store",
                "description": "Delete one store by it's id",
                "operationId": "deleteStore",
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "description": "ID of shop",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "shop": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Digitec"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "USB cable"
                    },
                    "image": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "977630ae369389ee38ca474af9e3d30e.png"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Products"
                }
            },
            "ShopResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Digitec"
                    },
                    "link": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "www.digitec.ch"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductResource"
                        },
                        "collectionFormat": "multi"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Shops"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}